Title: 이차원 배열의 오름차순 정리, Time: 0ms, Memory: 0MB, Status: Passed - Codetree

+++ 260401/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -3,15 +3,15 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int N = sc.nextInt();
-        int K = sc.nextInt();
+        long N = sc.nextLong();
+        long K = sc.nextLong();
 
-        int s = 1;
-        int e = N * N;
-        int answer = 0;
+        long s = 1;
+        long e = N * N;
+        long answer = 0;
         while (s <= e) {
-            int m = (s + e) / 2;
-            int count = 0;
+            long m = (s + e) / 2;
+            long count = 0;
 
             for (int i = 1; i <= N; i++) {
                 count += Math.min(N, m / i);
